backend/requirements.txt
fastapi
uvicorn
python-multipart
pillow
numpy
tensorflow
scikit-learn
matplotlib
seaborn
opencv-python
python-dotenv
pydantic
supabase
